pyproject.toml
[project]
name = "florianbrand.de"
version = "1.0.0"
description = "The FastHTML site of florianbrand.de"
requires-python = ">=3.12"
dependencies = [
    "fh-bootstrap>=0.0.3",
    "python-fasthtml>=0.10.0",
    "python-frontmatter>=1.1.0",
    "markdown-it-py[plugins]>=3.0.0",
    "python-fasthtml>=0.6.4",
    "feedgen>=1.0.0",
    "pytz>=2024.2",
    "pyyaml>=6.0.2",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
ignore = [
    "F403",
    "F405"
]
fixable = ["ALL"]
